# boardgameSorter/boardgameSorter/src/main/java/com/boardgame/boardgameSorter/entity/Mechanism.java
package com.boardgame.boardgameSorter.entity;

import jakarta.persistence.Entity;
import jakarta.persistence.GeneratedValue;
import jakarta.persistence.GenerationType;
import jakarta.persistence.Id;
import lombok.Getter;
import lombok.Setter;

@Setter
@Getter
@Entity
public class Mechanism {
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Integer id;

    private String mechanismName;

    private String mechanismDescription;
}